Make nodegit a dev dependency, hidden behind a config flag
Loading nodegit in every build is really slow, when we don't really use it in any real instance because we operate from inside a spotlight instance as a submodule.

Adding it as a dev dependency allows it to be available when running locally, but not incuded as part of the build process.